requirements.txt
fastapi
pydantic
httptools
psycopg2-binary
requests
SQLAlchemy
starlette
urllib3
uvicorn
uvicorn-logger
websocket-client
websockets
pandas
numpy
matplotlib
matplotlib-inline
scikit-learn
python-dateutil
tabulate
fpdf
openpyxl
hashlib_additional
pyyaml
python-json-logger
pyaml-env
pyotp
art
apscheduler
PyPortfolioOpt